lego: run_chroot: Enforce chroot to begin in HOME
This change is to enforce the working directory of a chroot.
to be HOME (/home/USERNAME). Some scripts and applications
expect the initial working directory to be explicitly declared.
By default, schroot enters a chroot at /home/USERNAME. Thus this
behavior is explicitly enforced to provide better defined behavior
for lego's interface.